+++ R/caseconverters.R
@@ -82,6 +82,7 @@
 #' to_swap_case(strings)
 #' to_sentence_case(strings)
 #' to_random_case(strings)
+#' to_title_case(strings)
 #' 
 #' 
 #' @rdname caseconverter
@@ -381,6 +382,34 @@ to_random_case <- function(string,
                          postfix = ""){
   to_any_case(string = string,
               case = "random",
+              sep_in = sep_in,
+              transliterations = transliterations,              
+              numerals = numerals,
+              sep_out = sep_out,
+              prefix = prefix,
+              postfix = postfix,
+              unique_sep = unique_sep,
+              empty_fill = empty_fill,
+              parsing_option = parsing_option,
+              abbreviations = abbreviations)
+}
+
+#' @rdname caseconverter
+#' @export
+
+to_title_case <- function(string,
+                         abbreviations = NULL,
+                         sep_in = "[^[:alnum:]]",
+                         parsing_option = 1,
+                         transliterations = NULL,             
+                         numerals = "middle",                   
+                         sep_out = NULL,
+                         unique_sep = NULL,
+                         empty_fill = NULL,
+                         prefix = "",
+                         postfix = ""){
+  to_any_case(string = string,
+              case = "title",
               sep_in = sep_in,
               transliterations = transliterations,              
               numerals = numerals,