+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * **numerals**: new alignment option `"tight"` which allows to suppress all underscores between numerals and letters.
 * **improved speed**: time of internal parsing could be reduced about a factor of ten, due to some `vapply()` replacements.
+* **UTF8 Conversion**: Input is now always converted to UTF-8 and returned as UTF8. Also the `transliterations` argument is now aware of non-UTF8 encoded input.
 
 # snakecase 0.9.2